requirements.txt
# Core Flask and Web Framework
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask-Migrate==4.1.0
Flask-Mail==0.10.0
Flask-Compress==1.18
Flask-Caching==2.3.0
Werkzeug==3.1.3

# Security and Cryptography
cryptography==41.0.7
bcrypt==4.1.2

# Database Drivers
psycopg==3.2.9
psycopg2-binary==2.9.9
PyMySQL==1.1.1
mysql-connector-python==9.4.0

# Production Server
gunicorn==23.0.0
gevent==23.9.1

# Image Processing
Pillow==10.4.0

# Data Processing and Export
openpyxl==3.1.5
pandas==2.3.1

# HTTP and Networking
requests==2.32.4
urllib3==2.2.2

# Configuration and Environment
python-dotenv==1.1.1

# Monitoring and Logging
sentry-sdk[flask]==1.45.0

# Rate Limiting
Flask-Limiter==3.5.0

# Session Management
Flask-Session==0.5.0
redis==5.0.1

# Development and Testing (optional)
pytest==7.4.4
pytest-flask==1.3.0
pytest-cov==4.1.0

# Code Quality (optional)
flake8==6.1.0
black==23.12.1

# Security Scanning (optional)
bandit==1.7.5
safety==2.3.5
